Added a pushpeg
Thinking about easier production, instead of hanging bits off the
pegboard, pin them in place.  This puts the marble module flat
against the pegboard, and is therefore incompatible with hangers.

The idea would be to transition everything to pegs - for most parts
that means the hanger can be shorter, but the ball return bits will
need some re-thinking.  Ideally these parts need to be easier to
remove and re-add...